+++ packages/models/src/box/sensorLayouts/sensorDefinitions/bme680_humidity.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  title: 'Luftfeuchte',
+  unit: '%',
+  sensorType: 'BME680',
+  icon: 'osem-humidity'
+};